Remove reference to first restart in Windows post & redirect
Basically, the post that was at:
http://robjwells.com/post/27373542696/restarting-in-bootcamp-the-easy-way

was my first go at it, and the information it contained was badly
wrong and unsafe. I don't like the idea of just pulling incorrect
posts but no-one should have followed the advice in there.

A newer post that linked to it has had the link removed (and that
post itself points to a more recent update on the same topic) and
in the nginx config file the Tumblr url for the original post
now redirects to the most recent post on the subject.

Not ideal, but cleaning up my mistakes from years ago.