Conditionally load JS if post or page contains code
Note that the 'conditional loading' is done at build time,
not at runtime in the browser.

A new extension marks each Post/Page object with the
include_highlighting attribute. If a single post/page has a
pre-code block then the JavaScript is included. If an index page
contains any posts with code then JavaScript is included.

Otherwise, the JavaScript is not included in the HTML file's head.