CSS: Adapt table styles to current browsers
Tables always have a width of 100%, and the default table-layout
setting (auto) will cause the table to expand past the container
if it’s too wide.

This is not exactly what I want, and not what my styles did when
I wrote them originally. My ideal is that table cells are sized to
the content, and if the table is too wide for its container then the
whole table will scroll horizontally, much in the way that pre-code
blocks do.

But I can’t seem to get that to work, and my styles caused each
table cell to scroll individually, which was an unreadable mess.

Width is 100% because tiny tables can look a bit scruffy in the
column of text, but 100% is not always what I want either as
tables fill the whole width of the container (while paragraphs
are limited to 35rem), but a consistent 100% is a better
compromise than trying to wrangle media queries that will
probably not do what I want all the time anyway.