Revert "Head: conditionally load highlight styles"
The saving for non-code pages was 0.5kB, while the penalty for code-containing pages was 1kB. At the moment, I have 78 pages without code, 41 pages with. That means that currently conditionally loading the syntax styles is a net loss (39kB - 41kB), and higher if we were to exclude non-blog pages.

It was a nice idea though!